txn: Move several MvccTxn functions to command (#8435)
