+++ src/storage/mvcc/txn.rs
@@ -25,7 +25,7 @@ pub struct GcInfo {
 
 /// Generate the Write record that should be written that means to to perform a specified rollback
 /// operation.
-fn make_rollback(
+pub(crate) fn make_rollback(
     start_ts: TimeStamp,
     protected: bool,
     overlapped_write: Option<Write>,
@@ -60,7 +60,7 @@ impl MissingLockAction {
         }
     }
 
-    fn rollback(rollback_if_not_exist: bool) -> MissingLockAction {
+    pub(crate) fn rollback(rollback_if_not_exist: bool) -> MissingLockAction {
         if rollback_if_not_exist {
             MissingLockAction::ProtectedRollback
         } else {
@@ -104,12 +104,12 @@ pub enum SecondaryLockStatus {
 }
 
 pub struct MvccTxn<S: Snapshot> {
-    reader: MvccReader<S>,
+    pub(crate) reader: MvccReader<S>,
     start_ts: TimeStamp,
     write_size: usize,
     writes: WriteData,
     // collapse continuous rollbacks.
-    collapse_rollback: bool,
+    pub(crate) collapse_rollback: bool,
     pub extra_op: ExtraOp,
     // `concurrency_manager` is used to set memory locks for prewritten keys.
     // Prewritten locks of async commit transactions should be visible to
@@ -208,13 +208,13 @@ impl<S: Snapshot> MvccTxn<S> {
         self.write_size
     }
 
-    fn put_lock(&mut self, key: Key, lock: &Lock) {
+    pub(crate) fn put_lock(&mut self, key: Key, lock: &Lock) {
         let write = Modify::Put(CF_LOCK, key, lock.to_bytes());
         self.write_size += write.size();
         self.writes.modifies.push(write);
     }
 
-    fn unlock_key(&mut self, key: Key, pessimistic: bool) -> Option<ReleasedLock> {
+    pub(crate) fn unlock_key(&mut self, key: Key, pessimistic: bool) -> Option<ReleasedLock> {
         let released = ReleasedLock::new(&key, pessimistic);
         let write = Modify::Delete(CF_LOCK, key);
         self.write_size += write.size();
@@ -234,7 +234,7 @@ impl<S: Snapshot> MvccTxn<S> {
         self.writes.modifies.push(write);
     }
 
-    fn put_write(&mut self, key: Key, ts: TimeStamp, value: Value) {
+    pub(crate) fn put_write(&mut self, key: Key, ts: TimeStamp, value: Value) {
         let write = Modify::Put(CF_WRITE, key.append_ts(ts), value);
         self.write_size += write.size();
         self.writes.modifies.push(write);
@@ -313,7 +313,7 @@ impl<S: Snapshot> MvccTxn<S> {
 
     // Check whether there's an overlapped write record, and then perform rollback. The actual behavior
     // to do the rollback differs according to whether there's an overlapped write record.
-    fn check_write_and_rollback_lock(
+    pub(crate) fn check_write_and_rollback_lock(
         &mut self,
         key: Key,
         lock: &Lock,
@@ -357,7 +357,12 @@ impl<S: Snapshot> MvccTxn<S> {
     /// break consistency. To solve the problem, add the timestamp of the current rollback to the
     /// lock. So when the lock is committed, it can check if it will overwrite a rollback record
     /// by checking the information in the lock.
-    fn mark_rollback_on_mismatching_lock(&mut self, key: &Key, mut lock: Lock, is_protected: bool) {
+    pub(crate) fn mark_rollback_on_mismatching_lock(
+        &mut self,
+        key: &Key,
+        mut lock: Lock,
+        is_protected: bool,
+    ) {
         assert_ne!(lock.ts, self.start_ts);
 
         if !is_protected {
@@ -905,7 +910,7 @@ impl<S: Snapshot> MvccTxn<S> {
         self.cleanup(key, TimeStamp::zero(), false)
     }
 
-    fn check_txn_status_missing_lock(
+    pub(crate) fn check_txn_status_missing_lock(
         &mut self,
         primary_key: Key,
         mismatch_lock: Option<Lock>,
@@ -1014,245 +1019,13 @@ impl<S: Snapshot> MvccTxn<S> {
         }
     }
 
-    /// Delete any pessimistic lock with small for_update_ts belongs to this transaction.
-    pub fn pessimistic_rollback(
-        &mut self,
-        key: Key,
-        for_update_ts: TimeStamp,
-    ) -> Result<Option<ReleasedLock>> {
-        fail_point!("pessimistic_rollback", |err| Err(make_txn_error(
-            err,
-            &key,
-            self.start_ts,
-        )
-        .into()));
-
-        if let Some(lock) = self.reader.load_lock(&key)? {
-            if lock.lock_type == LockType::Pessimistic
-                && lock.ts == self.start_ts
-                && lock.for_update_ts <= for_update_ts
-            {
-                return Ok(self.unlock_key(key, true));
-            }
-        }
-        Ok(None)
-    }
-
-    fn collapse_prev_rollback(&mut self, key: Key) -> Result<()> {
+    pub(crate) fn collapse_prev_rollback(&mut self, key: Key) -> Result<()> {
         if let Some((commit_ts, write)) = self.reader.seek_write(&key, self.start_ts)? {
             if write.write_type == WriteType::Rollback && !write.as_ref().is_protected() {
                 self.delete_write(key, commit_ts);
             }
         }
         Ok(())
-    }
-
-    /// Update a primary key's TTL if `advise_ttl > lock.ttl`.
-    ///
-    /// Returns the new TTL.
-    pub fn txn_heart_beat(&mut self, primary_key: Key, advise_ttl: u64) -> Result<u64> {
-        fail_point!("txn_heart_beat", |err| Err(make_txn_error(
-            err,
-            &primary_key,
-            self.start_ts,
-        )
-        .into()));
-
-        if let Some(mut lock) = self.reader.load_lock(&primary_key)? {
-            if lock.ts == self.start_ts {
-                if lock.ttl < advise_ttl {
-                    lock.ttl = advise_ttl;
-                    self.put_lock(primary_key, &lock);
-                } else {
-                    debug!(
-                        "txn_heart_beat with advise_ttl not large than current ttl";
-                        "primary_key" => %primary_key,
-                        "start_ts" => self.start_ts,
-                        "advise_ttl" => advise_ttl,
-                        "current_ttl" => lock.ttl,
-                    );
-                }
-                return Ok(lock.ttl);
-            }
-        }
-
-        debug!(
-            "txn_heart_beat invoked but lock is absent";
-            "primary_key" => %primary_key,
-            "start_ts" => self.start_ts,
-            "advise_ttl" => advise_ttl,
-        );
-        Err(ErrorInner::TxnLockNotFound {
-            start_ts: self.start_ts,
-            commit_ts: TimeStamp::zero(),
-            key: primary_key.into_raw()?,
-        }
-        .into())
-    }
-
-    /// Check the status of a transaction.
-    ///
-    /// This operation checks whether a transaction has expired its primary lock's TTL, rollback the
-    /// transaction if expired, or update the transaction's min_commit_ts according to the metadata
-    /// in the primary lock.
-    ///
-    /// When transaction T1 meets T2's lock, it may invoke this on T2's primary key. In this
-    /// situation, `self.start_ts` is T2's `start_ts`, `caller_start_ts` is T1's `start_ts`, and
-    /// the `current_ts` is literally the timestamp when this function is invoked. It may not be
-    /// accurate.
-    ///
-    /// Returns (`lock_ttl`, `commit_ts`, `is_pessimistic_txn`).
-    /// After checking, if the lock is still alive, it retrieves the Lock's TTL; if the transaction
-    /// is committed, get the commit_ts; otherwise, if the transaction is rolled back or there's
-    /// no information about the transaction, results will be both 0.
-    pub fn check_txn_status(
-        &mut self,
-        primary_key: Key,
-        caller_start_ts: TimeStamp,
-        current_ts: TimeStamp,
-        rollback_if_not_exist: bool,
-    ) -> Result<(TxnStatus, Option<ReleasedLock>)> {
-        fail_point!("check_txn_status", |err| Err(make_txn_error(
-            err,
-            &primary_key,
-            self.start_ts,
-        )
-        .into()));
-
-        match self.reader.load_lock(&primary_key)? {
-            Some(mut lock) if lock.ts == self.start_ts => {
-                if lock.use_async_commit && (!caller_start_ts.is_zero() || !current_ts.is_zero()) {
-                    return Err(ErrorInner::Other(box_err!(
-                        "cannot call check_txn_status with caller_start_ts or current_ts set on async commit transaction"
-                    ))
-                    .into());
-                }
-
-                let is_pessimistic_txn = !lock.for_update_ts.is_zero();
-
-                if lock.ts.physical() + lock.ttl < current_ts.physical() {
-                    assert!(!lock.use_async_commit);
-                    // If the lock is expired, clean it up.
-                    let released =
-                        self.check_write_and_rollback_lock(primary_key, &lock, is_pessimistic_txn)?;
-                    MVCC_CHECK_TXN_STATUS_COUNTER_VEC.rollback.inc();
-                    return Ok((TxnStatus::TtlExpire, released));
-                }
-
-                // If lock.min_commit_ts is 0, it's not a large transaction and we can't push forward
-                // its min_commit_ts otherwise the transaction can't be committed by old version TiDB
-                // during rolling update.
-                if !lock.min_commit_ts.is_zero()
-                    // If the caller_start_ts is max, it's a point get in the autocommit transaction.
-                    // We don't push forward lock's min_commit_ts and the point get can ingore the lock
-                    // next time because it's not committed.
-                    && !caller_start_ts.is_max()
-                    // Push forward the min_commit_ts so that reading won't be blocked by locks.
-                    && caller_start_ts >= lock.min_commit_ts
-                {
-                    assert!(!lock.use_async_commit);
-                    lock.min_commit_ts = caller_start_ts.next();
-
-                    if lock.min_commit_ts < current_ts {
-                        lock.min_commit_ts = current_ts;
-                    }
-
-                    self.put_lock(primary_key, &lock);
-                    MVCC_CHECK_TXN_STATUS_COUNTER_VEC.update_ts.inc();
-                }
-
-                Ok((
-                    TxnStatus::uncommitted(
-                        lock.ttl,
-                        lock.min_commit_ts,
-                        lock.use_async_commit,
-                        lock.secondaries,
-                    ),
-                    None,
-                ))
-            }
-            // The rollback must be protected, see more on
-            // [issue #7364](https://github.com/tikv/tikv/issues/7364)
-            l => self
-                .check_txn_status_missing_lock(
-                    primary_key,
-                    l,
-                    MissingLockAction::rollback(rollback_if_not_exist),
-                )
-                .map(|s| (s, None)),
-        }
-    }
-
-    /// Check the status of a secondary (optimistic) lock.
-    ///
-    /// It checks whether the given secondary lock exists. If the lock exists,
-    /// the lock information is returned. Otherwise, it searches the write CF
-    /// for the commit record of the lock and returns the commit timestamp
-    /// (0 if the lock is not committed).
-    ///
-    /// If the lock does not exist or is a pessimistic lock, to prevent the
-    /// status being changed, a rollback may be written and this rollback
-    /// needs to be protected.
-    pub fn check_secondary_lock(
-        &mut self,
-        key: &Key,
-        start_ts: TimeStamp,
-    ) -> Result<(SecondaryLockStatus, Option<ReleasedLock>)> {
-        let mut released_lock = None;
-        let mut mismatch_lock = None;
-        let (status, need_rollback, rollback_overlapped_write) =
-            match self.reader.load_lock(&key)? {
-                Some(lock) if lock.ts == start_ts => {
-                    if lock.lock_type == LockType::Pessimistic {
-                        released_lock = self.unlock_key(key.clone(), true);
-                        let overlapped_write = self
-                            .reader
-                            .get_txn_commit_record(&key, start_ts)?
-                            .unwrap_none();
-                        (SecondaryLockStatus::RolledBack, true, overlapped_write)
-                    } else {
-                        (SecondaryLockStatus::Locked(lock), false, None)
-                    }
-                }
-                l => {
-                    mismatch_lock = l;
-                    match self.reader.get_txn_commit_record(&key, start_ts)? {
-                        TxnCommitRecord::SingleRecord { commit_ts, write } => {
-                            let status = if write.write_type != WriteType::Rollback {
-                                SecondaryLockStatus::Committed(commit_ts)
-                            } else {
-                                SecondaryLockStatus::RolledBack
-                            };
-                            // We needn't write a rollback once there is a write record for it:
-                            // If it's a committed record, it cannot be changed.
-                            // If it's a rollback record, it either comes from another check_secondary_lock
-                            // (thus protected) or the client stops commit actively. So we don't need
-                            // to make it protected again.
-                            (status, false, None)
-                        }
-                        TxnCommitRecord::OverlappedRollback { .. } => {
-                            (SecondaryLockStatus::RolledBack, false, None)
-                        }
-                        TxnCommitRecord::None { overlapped_write } => {
-                            (SecondaryLockStatus::RolledBack, true, overlapped_write)
-                        }
-                    }
-                }
-            };
-        if need_rollback {
-            if let Some(l) = mismatch_lock {
-                self.mark_rollback_on_mismatching_lock(&key, l, true);
-            }
-            // We must protect this rollback in case this rollback is collapsed and a stale
-            // acquire_pessimistic_lock and prewrite succeed again.
-            if let Some(write) = make_rollback(start_ts, true, rollback_overlapped_write) {
-                self.put_write(key.clone(), start_ts, write.as_ref().to_bytes());
-                if self.collapse_rollback {
-                    self.collapse_prev_rollback(key.clone())?;
-                }
-            }
-        }
-        Ok((status, released_lock))
     }
 
     pub fn gc(&mut self, key: Key, safe_point: TimeStamp) -> Result<GcInfo> {
@@ -1378,7 +1151,7 @@ impl<S: Snapshot> fmt::Debug for MvccTxn<S> {
 }
 
 #[cfg(feature = "failpoints")]
-fn make_txn_error(s: Option<String>, key: &Key, start_ts: TimeStamp) -> ErrorInner {
+pub(crate) fn make_txn_error(s: Option<String>, key: &Key, start_ts: TimeStamp) -> ErrorInner {
     if let Some(s) = s {
         match s.to_ascii_lowercase().as_str() {
             "keyislocked" => {
@@ -1449,6 +1222,8 @@ mod tests {
     use crate::storage::kv::{Engine, RocksEngine, TestEngineBuilder};
     use crate::storage::mvcc::tests::*;
     use crate::storage::mvcc::{Error, ErrorInner, MvccReader};
+    use crate::storage::txn::commands::*;
+    use crate::storage::SecondaryLocksStatus;
     use kvproto::kvrpcpb::Context;
     use txn_types::{TimeStamp, SHORT_VALUE_MAX_LEN};
 
@@ -1740,9 +1515,9 @@ mod tests {
 
         must_prewrite_put(&engine, k, v, k, ts(10, 0));
         must_locked(&engine, k, ts(10, 0));
-        must_txn_heart_beat(&engine, k, ts(10, 0), 100, 100);
+        txn_heart_beat::tests::must_success(&engine, k, ts(10, 0), 100, 100);
         // Check the last txn_heart_beat has set the lock's TTL to 100.
-        must_txn_heart_beat(&engine, k, ts(10, 0), 90, 100);
+        txn_heart_beat::tests::must_success(&engine, k, ts(10, 0), 90, 100);
 
         // TTL not expired. Do nothing but returns an error.
         must_cleanup_err(&engine, k, ts(10, 0), ts(20, 0));
@@ -1841,7 +1616,7 @@ mod tests {
         let uncommitted = super::TxnStatus::uncommitted;
 
         must_prewrite_put_for_large_txn(&engine, k, v, k, ts(10, 0), 100, 0);
-        must_check_txn_status(
+        check_txn_status::tests::must_success(
             &engine,
             k,
             ts(10, 0),
@@ -1856,7 +1631,7 @@ mod tests {
         must_commit(&engine, k, ts(10, 0), ts(20, 1));
 
         must_prewrite_put_for_large_txn(&engine, k, v, k, ts(30, 0), 100, 0);
-        must_check_txn_status(
+        check_txn_status::tests::must_success(
             &engine,
             k,
             ts(30, 0),
@@ -1869,7 +1644,7 @@ mod tests {
 
         // If the min_commit_ts of the pessimistic lock is greater than prewrite's, use it.
         must_acquire_pessimistic_lock_for_large_txn(&engine, k, k, ts(60, 0), ts(60, 0), 100);
-        must_check_txn_status(
+        check_txn_status::tests::must_success(
             &engine,
             k,
             ts(60, 0),
@@ -2448,7 +2223,7 @@ mod tests {
         must_get_none(&engine, k, 28);
         // Currently we cannot avoid this.
         must_acquire_pessimistic_lock(&engine, k, k, 26, 29);
-        must_pessimistic_rollback(&engine, k, 26, 29);
+        pessimistic_rollback::tests::must_success(&engine, k, 26, 29);
         must_unlocked(&engine, k);
 
         // Non pessimistic key in pessimistic transaction.
@@ -2464,7 +2239,7 @@ mod tests {
         must_acquire_pessimistic_lock_err(&engine, k, k, 32, 32);
         // Currently we cannot avoid this.
         must_acquire_pessimistic_lock(&engine, k, k, 32, 34);
-        must_pessimistic_rollback(&engine, k, 32, 34);
+        pessimistic_rollback::tests::must_success(&engine, k, 32, 34);
         must_unlocked(&engine, k);
 
         // Acquire lock when there is lock with different for_update_ts.
@@ -2591,65 +2366,6 @@ mod tests {
     }
 
     #[test]
-    fn test_pessimistic_rollback() {
-        let engine = TestEngineBuilder::new().build().unwrap();
-
-        let k = b"k1";
-        let v = b"v1";
-
-        // Normal
-        must_acquire_pessimistic_lock(&engine, k, k, 1, 1);
-        must_pessimistic_locked(&engine, k, 1, 1);
-        must_pessimistic_rollback(&engine, k, 1, 1);
-        must_unlocked(&engine, k);
-        must_get_commit_ts_none(&engine, k, 1);
-        // Pessimistic rollback is idempotent
-        must_pessimistic_rollback(&engine, k, 1, 1);
-        must_unlocked(&engine, k);
-        must_get_commit_ts_none(&engine, k, 1);
-
-        // Succeed if the lock doesn't exist.
-        must_pessimistic_rollback(&engine, k, 2, 2);
-
-        // Do nothing if meets other transaction's pessimistic lock
-        must_acquire_pessimistic_lock(&engine, k, k, 2, 3);
-        must_pessimistic_rollback(&engine, k, 1, 1);
-        must_pessimistic_rollback(&engine, k, 1, 2);
-        must_pessimistic_rollback(&engine, k, 1, 3);
-        must_pessimistic_rollback(&engine, k, 1, 4);
-        must_pessimistic_rollback(&engine, k, 3, 3);
-        must_pessimistic_rollback(&engine, k, 4, 4);
-
-        // Succeed if for_update_ts is larger; do nothing if for_update_ts is smaller.
-        must_pessimistic_locked(&engine, k, 2, 3);
-        must_pessimistic_rollback(&engine, k, 2, 2);
-        must_pessimistic_locked(&engine, k, 2, 3);
-        must_pessimistic_rollback(&engine, k, 2, 4);
-        must_unlocked(&engine, k);
-
-        // Do nothing if rollbacks a non-pessimistic lock.
-        must_prewrite_put(&engine, k, v, k, 3);
-        must_locked(&engine, k, 3);
-        must_pessimistic_rollback(&engine, k, 3, 3);
-        must_locked(&engine, k, 3);
-
-        // Do nothing if meets other transaction's optimistic lock
-        must_pessimistic_rollback(&engine, k, 2, 2);
-        must_pessimistic_rollback(&engine, k, 2, 3);
-        must_pessimistic_rollback(&engine, k, 2, 4);
-        must_pessimistic_rollback(&engine, k, 4, 4);
-        must_locked(&engine, k, 3);
-
-        // Do nothing if committed
-        must_commit(&engine, k, 3, 4);
-        must_unlocked(&engine, k);
-        must_get_commit_ts(&engine, k, 3, 4);
-        must_pessimistic_rollback(&engine, k, 3, 3);
-        must_pessimistic_rollback(&engine, k, 3, 4);
-        must_pessimistic_rollback(&engine, k, 3, 5);
-    }
-
-    #[test]
     fn test_overwrite_pessimistic_lock() {
         let engine = TestEngineBuilder::new().build().unwrap();
 
@@ -2661,418 +2377,6 @@ mod tests {
         must_pessimistic_locked(&engine, k, 1, 2);
         must_acquire_pessimistic_lock(&engine, k, k, 1, 3);
         must_pessimistic_locked(&engine, k, 1, 3);
-    }
-
-    #[test]
-    fn test_txn_heart_beat() {
-        let engine = TestEngineBuilder::new().build().unwrap();
-
-        let (k, v) = (b"k1", b"v1");
-
-        let test = |ts| {
-            // Do nothing if advise_ttl is less smaller than current TTL.
-            must_txn_heart_beat(&engine, k, ts, 90, 100);
-            // Return the new TTL if the TTL when the TTL is updated.
-            must_txn_heart_beat(&engine, k, ts, 110, 110);
-            // The lock's TTL is updated and persisted into the db.
-            must_txn_heart_beat(&engine, k, ts, 90, 110);
-            // Heart beat another transaction's lock will lead to an error.
-            must_txn_heart_beat_err(&engine, k, ts - 1, 150);
-            must_txn_heart_beat_err(&engine, k, ts + 1, 150);
-            // The existing lock is not changed.
-            must_txn_heart_beat(&engine, k, ts, 90, 110);
-        };
-
-        // No lock.
-        must_txn_heart_beat_err(&engine, k, 5, 100);
-
-        // Create a lock with TTL=100.
-        // The initial TTL will be set to 0 after calling must_prewrite_put. Update it first.
-        must_prewrite_put(&engine, k, v, k, 5);
-        must_locked(&engine, k, 5);
-        must_txn_heart_beat(&engine, k, 5, 100, 100);
-
-        test(5);
-
-        must_locked(&engine, k, 5);
-        must_commit(&engine, k, 5, 10);
-        must_unlocked(&engine, k);
-
-        // No lock.
-        must_txn_heart_beat_err(&engine, k, 5, 100);
-        must_txn_heart_beat_err(&engine, k, 10, 100);
-
-        must_acquire_pessimistic_lock(&engine, k, k, 8, 15);
-        must_pessimistic_locked(&engine, k, 8, 15);
-        must_txn_heart_beat(&engine, k, 8, 100, 100);
-
-        test(8);
-
-        must_pessimistic_locked(&engine, k, 8, 15);
-    }
-
-    fn test_check_txn_status_impl(rollback_if_not_exist: bool) {
-        let engine = TestEngineBuilder::new().build().unwrap();
-
-        let (k, v) = (b"k1", b"v1");
-
-        let ts = TimeStamp::compose;
-
-        // Shortcuts
-        use super::TxnStatus::*;
-        let committed = TxnStatus::committed;
-        let uncommitted = TxnStatus::uncommitted;
-        let r = rollback_if_not_exist;
-
-        // Try to check a not exist thing.
-        if r {
-            must_check_txn_status(&engine, k, ts(3, 0), ts(3, 1), ts(3, 2), r, LockNotExist);
-            // A protected rollback record will be written.
-            must_get_rollback_protected(&engine, k, ts(3, 0), true);
-        } else {
-            must_check_txn_status_err(&engine, k, ts(3, 0), ts(3, 1), ts(3, 2), r);
-        }
-
-        // Lock the key with TTL=100.
-        must_prewrite_put_for_large_txn(&engine, k, v, k, ts(5, 0), 100, 0);
-        // The initial min_commit_ts is start_ts + 1.
-        must_large_txn_locked(&engine, k, ts(5, 0), 100, ts(5, 1), false);
-
-        // CheckTxnStatus with caller_start_ts = 0 and current_ts = 0 should just return the
-        // information of the lock without changing it.
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(5, 0),
-            0,
-            0,
-            r,
-            uncommitted(100, ts(5, 1), false, vec![]),
-        );
-
-        // Update min_commit_ts to current_ts.
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(5, 0),
-            ts(6, 0),
-            ts(7, 0),
-            r,
-            uncommitted(100, ts(7, 0), false, vec![]),
-        );
-        must_large_txn_locked(&engine, k, ts(5, 0), 100, ts(7, 0), false);
-
-        // Update min_commit_ts to caller_start_ts + 1 if current_ts < caller_start_ts.
-        // This case should be impossible. But if it happens, we prevents it.
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(5, 0),
-            ts(9, 0),
-            ts(8, 0),
-            r,
-            uncommitted(100, ts(9, 1), false, vec![]),
-        );
-        must_large_txn_locked(&engine, k, ts(5, 0), 100, ts(9, 1), false);
-
-        // caller_start_ts < lock.min_commit_ts < current_ts
-        // When caller_start_ts < lock.min_commit_ts, no need to update it.
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(5, 0),
-            ts(8, 0),
-            ts(10, 0),
-            r,
-            uncommitted(100, ts(9, 1), false, vec![]),
-        );
-        must_large_txn_locked(&engine, k, ts(5, 0), 100, ts(9, 1), false);
-
-        // current_ts < lock.min_commit_ts < caller_start_ts
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(5, 0),
-            ts(11, 0),
-            ts(9, 0),
-            r,
-            uncommitted(100, ts(11, 1), false, vec![]),
-        );
-        must_large_txn_locked(&engine, k, ts(5, 0), 100, ts(11, 1), false);
-
-        // For same caller_start_ts and current_ts, update min_commit_ts to caller_start_ts + 1
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(5, 0),
-            ts(12, 0),
-            ts(12, 0),
-            r,
-            uncommitted(100, ts(12, 1), false, vec![]),
-        );
-        must_large_txn_locked(&engine, k, ts(5, 0), 100, ts(12, 1), false);
-
-        // Logical time is also considered in the comparing
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(5, 0),
-            ts(13, 1),
-            ts(13, 3),
-            r,
-            uncommitted(100, ts(13, 3), false, vec![]),
-        );
-        must_large_txn_locked(&engine, k, ts(5, 0), 100, ts(13, 3), false);
-
-        must_commit(&engine, k, ts(5, 0), ts(15, 0));
-        must_unlocked(&engine, k);
-
-        // Check committed key will get the commit ts.
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(5, 0),
-            ts(12, 0),
-            ts(12, 0),
-            r,
-            committed(ts(15, 0)),
-        );
-        must_unlocked(&engine, k);
-
-        must_prewrite_put_for_large_txn(&engine, k, v, k, ts(20, 0), 100, 0);
-
-        // Check a committed transaction when there is another lock. Expect getting the commit ts.
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(5, 0),
-            ts(12, 0),
-            ts(12, 0),
-            r,
-            committed(ts(15, 0)),
-        );
-
-        // Check a not existing transaction, the result depends on whether `rollback_if_not_exist`
-        // is set.
-        if r {
-            must_check_txn_status(&engine, k, ts(6, 0), ts(12, 0), ts(12, 0), r, LockNotExist);
-            // And a rollback record will be written.
-            must_seek_write(
-                &engine,
-                k,
-                ts(6, 0),
-                ts(6, 0),
-                ts(6, 0),
-                WriteType::Rollback,
-            );
-        } else {
-            must_check_txn_status_err(&engine, k, ts(6, 0), ts(12, 0), ts(12, 0), r);
-        }
-
-        // TTL check is based on physical time (in ms). When logical time's difference is larger
-        // than TTL, the lock won't be resolved.
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(20, 0),
-            ts(21, 105),
-            ts(21, 105),
-            r,
-            uncommitted(100, ts(21, 106), false, vec![]),
-        );
-        must_large_txn_locked(&engine, k, ts(20, 0), 100, ts(21, 106), false);
-
-        // If physical time's difference exceeds TTL, lock will be resolved.
-        must_check_txn_status(&engine, k, ts(20, 0), ts(121, 0), ts(121, 0), r, TtlExpire);
-        must_unlocked(&engine, k);
-        must_seek_write(
-            &engine,
-            k,
-            TimeStamp::max(),
-            ts(20, 0),
-            ts(20, 0),
-            WriteType::Rollback,
-        );
-
-        // Push the min_commit_ts of pessimistic locks.
-        must_acquire_pessimistic_lock_for_large_txn(&engine, k, k, ts(4, 0), ts(130, 0), 200);
-        must_large_txn_locked(&engine, k, ts(4, 0), 200, ts(130, 1), true);
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(4, 0),
-            ts(135, 0),
-            ts(135, 0),
-            r,
-            uncommitted(200, ts(135, 1), false, vec![]),
-        );
-        must_large_txn_locked(&engine, k, ts(4, 0), 200, ts(135, 1), true);
-
-        // Commit the key.
-        must_pessimistic_prewrite_put(&engine, k, v, k, ts(4, 0), ts(130, 0), true);
-        must_commit(&engine, k, ts(4, 0), ts(140, 0));
-        must_unlocked(&engine, k);
-        must_get_commit_ts(&engine, k, ts(4, 0), ts(140, 0));
-
-        // Now the transactions are intersecting:
-        // T1: start_ts = 5, commit_ts = 15
-        // T2: start_ts = 20, rollback
-        // T3: start_ts = 4, commit_ts = 140
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(4, 0),
-            ts(10, 0),
-            ts(10, 0),
-            r,
-            committed(ts(140, 0)),
-        );
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(5, 0),
-            ts(10, 0),
-            ts(10, 0),
-            r,
-            committed(ts(15, 0)),
-        );
-        must_check_txn_status(&engine, k, ts(20, 0), ts(10, 0), ts(10, 0), r, RolledBack);
-
-        // Rollback expired pessimistic lock.
-        must_acquire_pessimistic_lock_for_large_txn(&engine, k, k, ts(150, 0), ts(150, 0), 100);
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(150, 0),
-            ts(160, 0),
-            ts(160, 0),
-            r,
-            uncommitted(100, ts(160, 1), false, vec![]),
-        );
-        must_large_txn_locked(&engine, k, ts(150, 0), 100, ts(160, 1), true);
-        must_check_txn_status(&engine, k, ts(150, 0), ts(160, 0), ts(260, 0), r, TtlExpire);
-        must_unlocked(&engine, k);
-        // Rolling back a pessimistic lock should leave Rollback mark.
-        must_seek_write(
-            &engine,
-            k,
-            TimeStamp::max(),
-            ts(150, 0),
-            ts(150, 0),
-            WriteType::Rollback,
-        );
-
-        // Rollback when current_ts is u64::max_value()
-        must_prewrite_put_for_large_txn(&engine, k, v, k, ts(270, 0), 100, 0);
-        must_large_txn_locked(&engine, k, ts(270, 0), 100, ts(270, 1), false);
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(270, 0),
-            ts(271, 0),
-            TimeStamp::max(),
-            r,
-            TtlExpire,
-        );
-        must_unlocked(&engine, k);
-        must_seek_write(
-            &engine,
-            k,
-            TimeStamp::max(),
-            ts(270, 0),
-            ts(270, 0),
-            WriteType::Rollback,
-        );
-
-        must_acquire_pessimistic_lock_for_large_txn(&engine, k, k, ts(280, 0), ts(280, 0), 100);
-        must_large_txn_locked(&engine, k, ts(280, 0), 100, ts(280, 1), true);
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(280, 0),
-            ts(281, 0),
-            TimeStamp::max(),
-            r,
-            TtlExpire,
-        );
-        must_unlocked(&engine, k);
-        must_seek_write(
-            &engine,
-            k,
-            TimeStamp::max(),
-            ts(280, 0),
-            ts(280, 0),
-            WriteType::Rollback,
-        );
-
-        // Don't push forward the min_commit_ts if the min_commit_ts of the lock is 0.
-        must_acquire_pessimistic_lock_with_ttl(&engine, k, k, ts(290, 0), ts(290, 0), 100);
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(290, 0),
-            ts(300, 0),
-            ts(300, 0),
-            r,
-            uncommitted(100, TimeStamp::zero(), false, vec![]),
-        );
-        must_large_txn_locked(&engine, k, ts(290, 0), 100, TimeStamp::zero(), true);
-        must_pessimistic_rollback(&engine, k, ts(290, 0), ts(290, 0));
-
-        must_prewrite_put_impl(
-            &engine,
-            k,
-            v,
-            k,
-            &None,
-            ts(300, 0),
-            false,
-            100,
-            TimeStamp::zero(),
-            1,
-            /* min_commit_ts */ TimeStamp::zero(),
-            false,
-        );
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(300, 0),
-            ts(310, 0),
-            ts(310, 0),
-            r,
-            uncommitted(100, TimeStamp::zero(), false, vec![]),
-        );
-        must_large_txn_locked(&engine, k, ts(300, 0), 100, TimeStamp::zero(), false);
-        must_rollback(&engine, k, ts(300, 0));
-
-        must_prewrite_put_for_large_txn(&engine, k, v, k, ts(310, 0), 100, 0);
-        must_large_txn_locked(&engine, k, ts(310, 0), 100, ts(310, 1), false);
-        // Don't push forward the min_commit_ts if caller_start_ts is max.
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(310, 0),
-            TimeStamp::max(),
-            ts(320, 0),
-            r,
-            uncommitted(100, ts(310, 1), false, vec![]),
-        );
-        must_commit(&engine, k, ts(310, 0), ts(315, 0));
-        must_check_txn_status(
-            &engine,
-            k,
-            ts(310, 0),
-            TimeStamp::max(),
-            ts(320, 0),
-            r,
-            committed(ts(315, 0)),
-        );
-    }
-
-    #[test]
-    fn test_check_txn_status() {
-        test_check_txn_status_impl(false);
-        test_check_txn_status_impl(true);
     }
 
     #[test]
@@ -3172,7 +2476,7 @@ mod tests {
             if *is_optimistic {
                 must_rollback(&engine, k, expected_lock_info.get_lock_version());
             } else {
-                must_pessimistic_rollback(
+                pessimistic_rollback::tests::must_success(
                     &engine,
                     k,
                     expected_lock_info.get_lock_version(),
@@ -3252,7 +2556,7 @@ mod tests {
             None
         );
         must_pessimistic_locked(&engine, k, 10, 10);
-        must_pessimistic_rollback(&engine, k, 10, 10);
+        pessimistic_rollback::tests::must_success(&engine, k, 10, 10);
 
         // Put
         must_prewrite_put(&engine, k, v, k, 10);
@@ -3272,7 +2576,7 @@ mod tests {
             Some(v.to_vec())
         );
         must_pessimistic_locked(&engine, k, 25, 25);
-        must_pessimistic_rollback(&engine, k, 25, 25);
+        pessimistic_rollback::tests::must_success(&engine, k, 25, 25);
 
         // Skip Write::Lock
         must_prewrite_lock(&engine, k, k, 30);
@@ -3282,7 +2586,7 @@ mod tests {
             Some(v.to_vec())
         );
         must_pessimistic_locked(&engine, k, 45, 45);
-        must_pessimistic_rollback(&engine, k, 45, 45);
+        pessimistic_rollback::tests::must_success(&engine, k, 45, 45);
 
         // Skip Write::Rollback
         must_rollback(&engine, k, 50);
@@ -3291,7 +2595,7 @@ mod tests {
             Some(v.to_vec())
         );
         must_pessimistic_locked(&engine, k, 55, 55);
-        must_pessimistic_rollback(&engine, k, 55, 55);
+        pessimistic_rollback::tests::must_success(&engine, k, 55, 55);
 
         // Delete
         must_prewrite_delete(&engine, k, k, 60);
@@ -3310,7 +2614,7 @@ mod tests {
             Some(v.to_vec())
         );
         must_pessimistic_locked(&engine, k, 75, 75);
-        must_pessimistic_rollback(&engine, k, 75, 75);
+        pessimistic_rollback::tests::must_success(&engine, k, 75, 75);
     }
 
     #[test]
@@ -3325,7 +2629,7 @@ mod tests {
             let for_update_ts = for_update_ts.into();
             must_acquire_pessimistic_lock(engine, key, key, start_ts, for_update_ts);
             // Delete the pessimistic lock to pretend write failure.
-            must_pessimistic_rollback(engine, key, start_ts, for_update_ts);
+            pessimistic_rollback::tests::must_success(engine, key, start_ts, for_update_ts);
         }
 
         let engine = TestEngineBuilder::new().build().unwrap();
@@ -3351,7 +2655,7 @@ mod tests {
         // KeyIsLocked; should fail.
         must_acquire_pessimistic_lock(&engine, k, k, 50, 50);
         must_pipelined_pessimistic_prewrite_put_err(&engine, k, &v, k, 60, 60, true);
-        must_pessimistic_rollback(&engine, k, 50, 50);
+        pessimistic_rollback::tests::must_success(&engine, k, 50, 50);
 
         // Pessimistic lock not exist and not pipelined; should fail.
         must_pessimistic_prewrite_put_err(&engine, k, &v, k, 70, 70, true);
@@ -3595,181 +2899,6 @@ mod tests {
     }
 
     #[test]
-    fn test_check_async_commit_txn_status() {
-        // The preparation work is the same as test_async_prewrite_primary.
-        let engine = TestEngineBuilder::new().build().unwrap();
-        let ctx = Context::default();
-
-        let snapshot = engine.snapshot(&ctx).unwrap();
-        let cm = ConcurrencyManager::new(42.into());
-        let mut txn = MvccTxn::new(snapshot, TimeStamp::new(2), true, cm.clone());
-
-        let mutation = Mutation::Put((Key::from_raw(b"key"), b"value".to_vec()));
-        txn.prewrite(
-            mutation,
-            b"key",
-            &Some(vec![b"key1".to_vec(), b"key2".to_vec(), b"key3".to_vec()]),
-            false,
-            0,
-            4,
-            TimeStamp::zero(),
-        )
-        .unwrap();
-        engine
-            .write(&ctx, WriteData::from_modifies(txn.into_modifies()))
-            .unwrap();
-
-        let do_check_txn_status = |rollback_if_not_exist| {
-            let snapshot = engine.snapshot(&ctx).unwrap();
-            let mut txn = MvccTxn::new(snapshot, TimeStamp::new(2), true, cm.clone());
-            let (txn_status, released_lock) = txn
-                .check_txn_status(
-                    Key::from_raw(b"key"),
-                    0.into(),
-                    0.into(),
-                    rollback_if_not_exist,
-                )
-                .unwrap();
-            assert_eq!(
-                txn_status,
-                TxnStatus::uncommitted(
-                    0,
-                    43.into(), // min_commit_ts calculated from max_read_ts
-                    true,
-                    vec![b"key1".to_vec(), b"key2".to_vec(), b"key3".to_vec()]
-                )
-            );
-            assert!(released_lock.is_none());
-        };
-        do_check_txn_status(true);
-        do_check_txn_status(false);
-
-        // Disallow calling check_txn_status on async commit transactions with caller_start_ts or
-        // current_ts set.
-        must_check_txn_status_err(&engine, b"key", 2, 1, 0, true);
-        must_check_txn_status_err(&engine, b"key", 2, 0, 1, true);
-    }
-
-    #[test]
-    fn test_check_async_commit_secondary_locks() {
-        let engine = TestEngineBuilder::new().build().unwrap();
-        let ctx = Context::default();
-        let cm = ConcurrencyManager::new(1.into());
-
-        let check_secondary = |key, ts| {
-            let snapshot = engine.snapshot(&ctx).unwrap();
-            let key = Key::from_raw(key);
-            let ts = TimeStamp::new(ts);
-            let mut txn = MvccTxn::new(snapshot, ts, true, cm.clone());
-            let res = txn.check_secondary_lock(&key, ts).unwrap();
-            let modifies = txn.into_modifies();
-            if !modifies.is_empty() {
-                engine
-                    .write(&ctx, WriteData::from_modifies(modifies))
-                    .unwrap();
-            }
-            res
-        };
-
-        must_prewrite_lock(&engine, b"k1", b"key", 1);
-        must_commit(&engine, b"k1", 1, 3);
-        must_rollback(&engine, b"k1", 5);
-        must_prewrite_lock(&engine, b"k1", b"key", 7);
-        must_commit(&engine, b"k1", 7, 9);
-
-        // Lock CF has no lock
-        //
-        // LOCK CF       | WRITE CF
-        // --------------+---------------------
-        //               | 9: start_ts = 7
-        //               | 5: rollback
-        //               | 3: start_ts = 1
-
-        assert_eq!(
-            check_secondary(b"k1", 7),
-            (SecondaryLockStatus::Committed(9.into()), None)
-        );
-        must_get_commit_ts(&engine, b"k1", 7, 9);
-        assert_eq!(
-            check_secondary(b"k1", 5),
-            (SecondaryLockStatus::RolledBack, None)
-        );
-        must_get_rollback_ts(&engine, b"k1", 5);
-        assert_eq!(
-            check_secondary(b"k1", 1),
-            (SecondaryLockStatus::Committed(3.into()), None)
-        );
-        must_get_commit_ts(&engine, b"k1", 1, 3);
-        assert_eq!(
-            check_secondary(b"k1", 6),
-            (SecondaryLockStatus::RolledBack, None)
-        );
-        must_get_rollback_protected(&engine, b"k1", 6, true);
-
-        // ----------------------------
-
-        must_acquire_pessimistic_lock(&engine, b"k1", b"key", 11, 11);
-
-        // Lock CF has a pessimistic lock
-        //
-        // LOCK CF       | WRITE CF
-        // ------------------------------------
-        // ts = 11 (pes) | 9: start_ts = 7
-        //               | 5: rollback
-        //               | 3: start_ts = 1
-
-        let (status, released_lock) = check_secondary(b"k1", 11);
-        assert_eq!(status, SecondaryLockStatus::RolledBack);
-        assert!(released_lock.unwrap().pessimistic);
-        must_get_rollback_protected(&engine, b"k1", 11, true);
-
-        // ----------------------------
-
-        must_prewrite_lock(&engine, b"k1", b"key", 13);
-
-        // Lock CF has an optimistic lock
-        //
-        // LOCK CF       | WRITE CF
-        // ------------------------------------
-        // ts = 13 (opt) | 11: rollback
-        //               |  9: start_ts = 7
-        //               |  5: rollback
-        //               |  3: start_ts = 1
-
-        match check_secondary(b"k1", 13) {
-            (SecondaryLockStatus::Locked(_), None) => {}
-            res => panic!("unexpected lock status: {:?}", res),
-        }
-        must_locked(&engine, b"k1", 13);
-
-        // ----------------------------
-
-        must_commit(&engine, b"k1", 13, 15);
-
-        // Lock CF has an optimistic lock
-        //
-        // LOCK CF       | WRITE CF
-        // ------------------------------------
-        //               | 15: start_ts = 13
-        //               | 11: rollback
-        //               |  9: start_ts = 7
-        //               |  5: rollback
-        //               |  3: start_ts = 1
-
-        match check_secondary(b"k1", 14) {
-            (SecondaryLockStatus::RolledBack, None) => {}
-            res => panic!("unexpected lock status: {:?}", res),
-        }
-        must_get_rollback_protected(&engine, b"k1", 14, true);
-
-        match check_secondary(b"k1", 15) {
-            (SecondaryLockStatus::RolledBack, None) => {}
-            res => panic!("unexpected lock status: {:?}", res),
-        }
-        must_get_overlapped_rollback(&engine, b"k1", 15, 13, WriteType::Lock);
-    }
-
-    #[test]
     fn test_txn_timestamp_overlapping() {
         let engine = TestEngineBuilder::new().build().unwrap();
         let (k, v) = (b"k1", b"v1");
@@ -3832,13 +2961,18 @@ mod tests {
         assert!(w.has_overlapped_rollback);
 
         must_prewrite_put_async_commit(&engine, k, v, k, &Some(vec![]), 20, 0);
-        must_check_txn_status(&engine, k, 25, 0, 0, true, TxnStatus::LockNotExist);
+        check_txn_status::tests::must_success(&engine, k, 25, 0, 0, true, TxnStatus::LockNotExist);
         must_commit(&engine, k, 20, 25);
         let w = must_written(&engine, k, 20, 25, WriteType::Put);
         assert!(w.has_overlapped_rollback);
 
         must_prewrite_put_async_commit(&engine, k, v, k, &Some(vec![]), 30, 0);
-        must_check_secondary_lock(&engine, k, 35, SecondaryLockStatus::RolledBack);
+        check_secondary_locks::tests::must_success(
+            &engine,
+            k,
+            35,
+            SecondaryLocksStatus::RolledBack,
+        );
         must_commit(&engine, k, 30, 35);
         let w = must_written(&engine, k, 30, 35, WriteType::Put);
         assert!(w.has_overlapped_rollback);
