+++ src/storage/txn/commands/pessimistic_rollback.rs
@@ -2,13 +2,14 @@
 
 use crate::storage::kv::WriteData;
 use crate::storage::lock_manager::LockManager;
-use crate::storage::mvcc::MvccTxn;
+use crate::storage::mvcc::{MvccTxn, Result as MvccResult};
 use crate::storage::txn::commands::{
     Command, CommandExt, ReleasedLocks, TypedCommand, WriteCommand, WriteContext, WriteResult,
 };
 use crate::storage::txn::Result;
 use crate::storage::{ProcessResult, Result as StorageResult, Snapshot};
-use txn_types::{Key, TimeStamp};
+use std::mem;
+use txn_types::{Key, LockType, TimeStamp};
 
 command! {
     /// Rollback pessimistic locks identified by `start_ts` and `for_update_ts`.
@@ -35,7 +36,8 @@ impl CommandExt for PessimisticRollback {
 }
 
 impl<S: Snapshot, L: LockManager> WriteCommand<S, L> for PessimisticRollback {
-    fn process_write(self, snapshot: S, context: WriteContext<'_, L>) -> Result<WriteResult> {
+    /// Delete any pessimistic lock with small for_update_ts belongs to this transaction.
+    fn process_write(mut self, snapshot: S, context: WriteContext<'_, L>) -> Result<WriteResult> {
         let mut txn = MvccTxn::new(
             snapshot,
             self.start_ts,
@@ -43,22 +45,147 @@ impl<S: Snapshot, L: LockManager> WriteCommand<S, L> for PessimisticRollback {
             context.concurrency_manager,
         );
 
-        let rows = self.keys.len();
+        let ctx = mem::take(&mut self.ctx);
+        let keys = mem::take(&mut self.keys);
+
+        let rows = keys.len();
         let mut released_locks = ReleasedLocks::new(self.start_ts, TimeStamp::zero());
-        for k in self.keys {
-            released_locks.push(txn.pessimistic_rollback(k, self.for_update_ts)?);
+        for key in keys {
+            fail_point!("pessimistic_rollback", |err| Err(
+                crate::storage::mvcc::Error::from(crate::storage::mvcc::txn::make_txn_error(
+                    err,
+                    &key,
+                    self.start_ts
+                ))
+                .into()
+            ));
+            let released_lock: MvccResult<_> = if let Some(lock) = txn.reader.load_lock(&key)? {
+                if lock.lock_type == LockType::Pessimistic
+                    && lock.ts == self.start_ts
+                    && lock.for_update_ts <= self.for_update_ts
+                {
+                    Ok(txn.unlock_key(key, true))
+                } else {
+                    Ok(None)
+                }
+            } else {
+                Ok(None)
+            };
+            released_locks.push(released_lock?);
         }
         released_locks.wake_up(context.lock_mgr);
 
         context.statistics.add(&txn.take_statistics());
         let write_data = WriteData::from_modifies(txn.into_modifies());
         Ok(WriteResult {
-            ctx: self.ctx,
+            ctx,
             to_be_write: write_data,
             rows,
             pr: ProcessResult::MultiRes { results: vec![] },
             lock_info: None,
             lock_guards: vec![],
         })
+    }
+}
+
+#[cfg(test)]
+pub mod tests {
+    use super::*;
+    use crate::storage::kv::Engine;
+    use crate::storage::lock_manager::DummyLockManager;
+    use crate::storage::mvcc::tests::*;
+    use crate::storage::txn::commands::{WriteCommand, WriteContext};
+    use crate::storage::TestEngineBuilder;
+    use concurrency_manager::ConcurrencyManager;
+    use kvproto::kvrpcpb::Context;
+    use txn_types::Key;
+
+    pub fn must_success<E: Engine>(
+        engine: &E,
+        key: &[u8],
+        start_ts: impl Into<TimeStamp>,
+        for_update_ts: impl Into<TimeStamp>,
+    ) {
+        let ctx = Context::default();
+        let snapshot = engine.snapshot(&ctx).unwrap();
+        let for_update_ts = for_update_ts.into();
+        let cm = ConcurrencyManager::new(for_update_ts);
+        let start_ts = start_ts.into();
+        let command = crate::storage::txn::commands::PessimisticRollback {
+            ctx: ctx.clone(),
+            keys: vec![Key::from_raw(key)],
+            start_ts,
+            for_update_ts,
+        };
+        let lock_mgr = DummyLockManager;
+        let write_context = WriteContext {
+            lock_mgr: &lock_mgr,
+            concurrency_manager: cm,
+            extra_op: Default::default(),
+            statistics: &mut Default::default(),
+            pipelined_pessimistic_lock: false,
+            enable_async_commit: true,
+        };
+        let result = command.process_write(snapshot, write_context).unwrap();
+        write(engine, &ctx, result.to_be_write.modifies);
+    }
+
+    #[test]
+    fn test_pessimistic_rollback() {
+        let engine = TestEngineBuilder::new().build().unwrap();
+
+        let k = b"k1";
+        let v = b"v1";
+
+        // Normal
+        must_acquire_pessimistic_lock(&engine, k, k, 1, 1);
+        must_pessimistic_locked(&engine, k, 1, 1);
+        must_success(&engine, k, 1, 1);
+        must_unlocked(&engine, k);
+        must_get_commit_ts_none(&engine, k, 1);
+        // Pessimistic rollback is idempotent
+        must_success(&engine, k, 1, 1);
+        must_unlocked(&engine, k);
+        must_get_commit_ts_none(&engine, k, 1);
+
+        // Succeed if the lock doesn't exist.
+        must_success(&engine, k, 2, 2);
+
+        // Do nothing if meets other transaction's pessimistic lock
+        must_acquire_pessimistic_lock(&engine, k, k, 2, 3);
+        must_success(&engine, k, 1, 1);
+        must_success(&engine, k, 1, 2);
+        must_success(&engine, k, 1, 3);
+        must_success(&engine, k, 1, 4);
+        must_success(&engine, k, 3, 3);
+        must_success(&engine, k, 4, 4);
+
+        // Succeed if for_update_ts is larger; do nothing if for_update_ts is smaller.
+        must_pessimistic_locked(&engine, k, 2, 3);
+        must_success(&engine, k, 2, 2);
+        must_pessimistic_locked(&engine, k, 2, 3);
+        must_success(&engine, k, 2, 4);
+        must_unlocked(&engine, k);
+
+        // Do nothing if rollbacks a non-pessimistic lock.
+        must_prewrite_put(&engine, k, v, k, 3);
+        must_locked(&engine, k, 3);
+        must_success(&engine, k, 3, 3);
+        must_locked(&engine, k, 3);
+
+        // Do nothing if meets other transaction's optimistic lock
+        must_success(&engine, k, 2, 2);
+        must_success(&engine, k, 2, 3);
+        must_success(&engine, k, 2, 4);
+        must_success(&engine, k, 4, 4);
+        must_locked(&engine, k, 3);
+
+        // Do nothing if committed
+        must_commit(&engine, k, 3, 4);
+        must_unlocked(&engine, k);
+        must_get_commit_ts(&engine, k, 3, 4);
+        must_success(&engine, k, 3, 3);
+        must_success(&engine, k, 3, 4);
+        must_success(&engine, k, 3, 5);
     }
 }
